Reorganized peripherals into their own module
