+++ .gitignore
@@ -1,5 +1,7 @@
 # IntellJ IDEA files
 .idea
+# VSCode files
+/.vscode/
 
 # Generated by Cargo
 # will have compiled files and executables
