+++ src/peripherals/simple_peripheral.rs
@@ -1,5 +1,5 @@
 use std::collections::HashMap;
-use crate::peripheral::{Peripheral, MethodsMap};
+use super::{Peripheral, MethodsMap};
 
 struct SimplePeripheral {}
